Непонятки с энкодером

+++ VirtualComPort/Protocol/can.h
@@ -13,7 +13,7 @@ typedef struct CRYPT_MSG
 
 
 extern CRYPT_MSG receiveMsg;
-void encryptTxMsg(CanTxMsgTypeDef *pTxMsg, uint32_t idSender, uint32_t idRecipient, uint32_t command, 
+void encryptTxMsg(CanTxMsgTypeDef *pTxMsg, uint8_t idSender, uint8_t idRecipient, uint8_t command, 
                 uint8_t commandTP, uint16_t address, uint32_t parameter);
 				
 void decryptRxMsg(CanRxMsgTypeDef *pRxMsg, CRYPT_MSG *pCryptMsg);
